Several fixes on Log, Protocol, Restart, Add Reschedule task
git-svn-id: http://goldengate.svn.sourceforge.net/svnroot/goldengate/GoldenGateR66/trunk@827 7864f6d1-e753-4b29-85b9-74ae93846b71

+++ src/main/java/openr66/client/SendThroughClient.java
@@ -204,6 +204,7 @@ public abstract class SendThroughClient extends AbstractTransfer {
                 }
             }
             if (exc!= null) {
+                taskRunner.setLocalChannelReference(new LocalChannelReference());
                 logger.error("Cannot Connect", exc);
                 future.setResult(new R66Result(exc, null, true,
                         ErrorCode.ConnectionImpossible, taskRunner));
@@ -279,7 +280,7 @@ public abstract class SendThroughClient extends AbstractTransfer {
             }
         } finally {
             if (taskRunner != null) {
-                if ((!future.isSuccess()) || nolog) {
+                if ((future.isDone() && (!future.isSuccess())) || nolog) {
                     try {
                         taskRunner.delete();
                     } catch (GoldenGateDatabaseException e) {
